fix: remove request for svg output
