+++ app/core/resources/schemas/enums/image_type_enum.py
@@ -15,5 +15,3 @@ class ImageTypeEnum(str, Enum):
     PNG = "png"
 
     GIF = "gif"
-
-    SVG = "svg"
