+++ package/preview/messages.ini
@@ -19,4 +19,3 @@ format_not_supported_error = Format not supported.
 file_not_valid_error = The input file should not be null.
 document_thumbnail_not_enabled_error = The document thumbnail function is not currently enabled!
 document_preview_not_enabled_error = The document preview function is not currently enabled!
-conversion_not_supported_error = The conversion from original image type to requested type is not supported!
