+++ rest-api.yaml
@@ -1027,7 +1027,6 @@ components:
         - jpeg
         - png
         - gif
-        - svg
       type: string
       description: |2-
